+++ src/settings/StatisticalCodeSettings.test.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { render, screen } from '@testing-library/react';
+import { render, screen } from '@folio/jest-config-stripes/testing-library/react';
 import { Form } from 'react-final-form';
 import { IntlProvider } from 'react-intl';
 import '../../test/jest/__mock__';