+++ CHANGELOG.md
@@ -17,6 +17,7 @@
 * Chronology not displayed in receiving history. Fixes UIIN-2411.
 * Also support `circulation` `14.0`. Refs UIIN-2412.
 * Instance/Holdings/Item notes, administrative notes character limit to 32K. Refs UIIN-2354.
+* Import testing-library deps from `@folio/jest-config-stripes`. Refs UIIN-2427.
 
 ## [9.4.5](https://github.com/folio-org/ui-inventory/tree/v9.4.5) (2023-04-03)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.4.4...v9.4.5)
@@ -132,7 +133,7 @@
 * Added `<ControllableDetail>` component to display authorized indicator for Subjects, Uniform titles and Series. Refs UIIN-2311.
 * Moved the print button from QuickMarkView, into the ViewSource. Due to the folio-org/ui-quick-marc#468. Refs UIIN-2324.
 * Change title for the print popup. Refs UIIN-2329.
-* Move @testing-library/* to dev-deps. Refs UIIN-2309.
+* Move @folio/jest-config-stripes/testing-library/* to dev-deps. Refs UIIN-2309.
 
 ## [9.2.0](https://github.com/folio-org/ui-inventory/tree/v9.2.0) (2022-10-27)
 [Full Changelog](https://github.com/folio-org/ui-inventory/compare/v9.1.0...v9.2.0)